DWH_1_Detecting_novel_associations_in_large_data_sets/requirements.txt
numpy>=1.21
pandas>=1.3
scikit-learn>=1.0
matplotlib>=3.4
pytest>=7.0
# Optional (not required):
# minepy>=1.0
python-pptx>=0.6.21
